chore: seperating trained model into its own file
